Add-wishList-structure
