+++ frontend/src/App.css
@@ -8,6 +8,7 @@
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
+  background-color: #f1eee3;
  
 }
