--- frontend/src/components/WishListItem.jsx
@@ -1,4 +1,0 @@
-const WishListItem = () => {
-  return <div>WishListItem</div>
-}
-export default WishListItem
