+++ frontend/src/components/wishList/WishEmpty.jsx
@@ -1,12 +1,12 @@
 const WishEmpty = () => {
   return (
-    <>
-      <p>your Wishlist is empty!</p>
+    <article>
+      <p className="mb-4 font-semibold">your Wishlist is empty!</p>
       <p>
         You don't have any products added in your wishlist. Search and save
         items to your liking!
       </p>
-    </>
+    </article>
   )
 }
 export default WishEmpty
